Updated all classes.
Originally, button creation was handled by the View class. In this revision, the task of button creation has been moved to the Controller class. Controller class creates an TicTacToeButton[] array and stores the created buttons in the array. View class displays these buttons by using a for-loop to iterate through the TTTButton array and display the btn at index 'i' of the array.